% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{adjust_park_factors}
\alias{adjust_park_factors}
\title{Adjust statistics for park effects}
\usage{
adjust_park_factors(stats, pfs, type = c("bat", "pit"), b_bio = NULL)
}
\arguments{
\item{stats}{data.frame of player statistics.  Obtained from
\code{\link{clean_PIA}}}

\item{pfs}{data.frame of multipliers for park factors.  Must be in proper
form.}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting.}

\item{b_bio}{optional data.frame.  Contains the handedness of batters.  Used
because park factors vary for left vs right handed batters.}
}
\value{
\code{tbl_df} of statistics that have been adjusted for park factors.
}
\description{
This takes a data.frame returned from \code{\link{clean_PIA}} and adjusts the
statistics for park effects.
}
\details{
If \code{type = "bat"}, then \code{stats} is left joined with
  \code{b_bio} and the two are joined by the \code{MLBID} column.  This will
  add the \code{Bats} column to the \code{stats} data.  For pitchers, the
  \code{Bats} column is set to be "Both" (i.e. they are treated as switch
  hitters).

  Next, the statistics are turned into long format using \code{\link{gather}}
  from the \code{tidyr} package.  If any park factors are missing,
  they are set to average (which is 100).  Statistics are then adjusted using
  the formula: \deqn{adjusted = count / (PF / 200 + .5)} Extraneous columns
  are discarded then the data are returned to a wide format using
  \code{\link{spread}} from the \code{tidyr} package.
}
\examples{
curr_wd <- getwd()
setwd("N:/Apps/simScoresApp/data")
stats <- read.csv("1-cleaned/batters/pia.csv", header = T, stringsAsFactors = F)
pfs <- read.csv("manual-info/Park_Factors.csv", header = T, stringsAsFactors = F)
b_bio <- read.csv("manual-info/bio_bat.csv", header = T, stringsAsFactors = F)
x <- adjust_park_factors(stats, pfs, "bat", b_bio)
setwd(curr_wd)
}

