% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{adjust_year_effects}
\alias{adjust_year_effects}
\title{Adjust for Changes in Run-Scoring Environment}
\usage{
adjust_year_effects(stats, yfs, type = c("bat", "pit"))
}
\arguments{
\item{stats}{data.frame of player statistics.  Obtained from
\code{\link{adjust_park_factors}}.}

\item{yfs}{data.frame of multipliers for year effects.  Must be in proper
form.}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting}
}
\value{
\code{tbl_df} of statistics that have been adjusted for year effects.
}
\description{
Uses year multipliers to adjust statistics to account for changes in run
scoring environment.  This is particularly useful because a 0.300 OBP in 2006
is much different from a 0.300 OBP in 2014.
}
\details{
\code{stats} are left joined with \code{yfs}.  If any year factors
  are missing, they are set to average (which is 100).  The statistics are
  turned into long format using \code{\link{gather}} from the
  \code{tidyr} package.  Statistics are then adjusted using the
  formula: \deqn{adjusted = Count * 100 / YF} Extraneous columns are
  discarded then the data are returned to a wide format using
  \code{\link{spread}} from the \code{tidyr} package.
}
\examples{
curr_wd <- getwd()
setwd("N:/Apps/simScoresApp/data")
stats <- read.csv("2-park-adjusted/bat-pf-adjust.csv", header = T, stringsAsFactors = F)
yfs <- read.csv("manual-info/Year-Factors.csv", header = T, stringsAsFactors = F)
x <- adjust_year_effects(stats, yfs, type = "bat")
setwd(curr_wd)
}

