% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarity-scores.R
\name{age_comps}
\alias{age_comps}
\title{Use MLEs to Calculate Similarity Scores for Batters}
\usage{
age_comps(player, data, weights, years = 3, type = c("bat", "pit"))
}
\arguments{
\item{player}{character.  The name of the focus player, formatted as "Last;
First M."}

\item{data}{The dataset of MLEs.  Should be the \code{level_grp} dataset
from `\code{simScoresApp/b-comparison-app/bat_simscore_data.RData}.}

\item{weights}{Numeric, named vector.  The names tell which variables to use
for comparisons.  The value tells the weight put on that variable when
comparing.}

\item{years}{numeric vector of length 1 or 2.  If length 1, specifies how
many years of statistcs to include.  If length two, specifies the age range
over which to compare (inclusive).}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting.}
}
\description{
Use MLEs to Calculate Similarity Scores for Batters
}
\examples{
load("H:/simScoresApp/b-comparison-app/bat_simscore_data.RData")
wts <- c(position = .05, pa = .9, bb.perc = .66, k.perc = .84, iso = .84, babip = .23, obp = .92)
# comparing stats based on three most recent years
age_comps("Brown; Domonic L.", age_grp, wts, 3, "bat")
# comparing stats from age 24 to 26
age_comps("Brown; Domonic L.", age_grp, wts, c(24, 26), "bat")
# same thing!
}

