% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{calc_MLEs}
\alias{calc_MLEs}
\title{Calculates Major League Equivalencies (MLEs) for specified statistics}
\usage{
calc_MLEs(stats, mles, type = c("bat", "pit"))
}
\arguments{
\item{stats}{data.frame of player statistics.  Obtained from
adjust_park_factors()}

\item{mles}{data.frame of multipliers for MLEs.  Must be in proper form.}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting}
}
\value{
\code{tbl_df} of statistics that have been adjusted for level
  effects.
}
\description{
Takes statistics returned from adjust_park_factors() and turns them into MLEs
}
\details{
\code{stats} are left joined with \code{mles}.  If any MLEs are
  missing, they are set to average (which is 1).  The statistics are turned
  into long format using \code{\link{gather}} from the \code{tidyr}
  package.

  For batters statistics are adjusted by multiplying by the multiplier
  \deqn{adjusted = Count * MLE}.  For pitchers, statistics are divided by the
  multiplier \deqn{adjusted = Count / MLE}. Extraneous columns are discarded
  then the data are returned to a wide format using \code{\link{spread}} from
  the \code{tidyr} package.

  Since the same multiplier is used for batters and pitchers, it is assumed
  that the transition from one level to the next is the same for batters and
  pitchers.  That is, this assumes that strikeouts increase for batters and
  decrease for pitchers at the same rate when changing level.  Additional
  research could be done to see if this assumption is valid.
}
\examples{
curr_sd <- getwd()
setwd("N:/Apps/simScoresApp/data")
stats <- read.csv("3-year-adjusted/bat-yf-adjust.csv", header = T, stringsAsFactors = F)
mults <- read.csv("manual-info/Level_Multipliers.csv", header = T, stringsAsFactors = F)
x <- calc_MLEs(stats, mults, type = "bat")
setwd(curr_wd)
}

