% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarity-scores.R
\name{calcScores2}
\alias{calcScores2}
\title{Wrapper to Calculate Similarity Scores}
\usage{
calcScores2(dat, player, weights)
}
\arguments{
\item{dat}{data.frame of player information.  Must be in proper format.}

\item{player}{name of player to find similarity scores for.}

\item{weights}{named vector.  The names are the stats for which we will be
calculating similarity scores.  The values are the weight for each
statistics to use when calculating similarity scores.}
}
\value{
data.frame containing scores for each player
}
\description{
Uses the \code{\link{findDiffs2}} function to calculate similarity scores.
This is used internally by \code{\link{age_comps}}, and
\code{\link{level_comps}}.  This function should only be used inside these
two functions.
}
\references{
Gower, J. C. (1971) A general coefficient of similarity and some
  of its properties, Biometrics 27, 857 - 874.
}

