% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarity-scores.R
\name{catDiffs2}
\alias{catDiffs2}
\title{Differences in Categorical Variables}
\usage{
catDiffs2(vec, dat)
}
\arguments{
\item{vec}{a character vector.  Each value in the vector corresponds to a
unique variable.}

\item{dat}{a character matrix or data.frame of character variables.  The
number of columns must be equal to the length of \code{vec}.}
}
\value{
a matrix of the same dimensions as \code{dat}.  The ith entry in the
  jth row is 0 if the jth value of \code{vec} is equal to the jth value of
  the ith row of {dat}.  It is 1 if they are different.  If either the ith
  value of the jth row of \code{dat} or the jth value of \code{vec} are
  missing, and NA will be returned.
}
\description{
Computes the scaled difference between a vector of categorical variables and
a matrix of categorical variables.  Should not usually be called by hand.
Instead, use the \code{\link{calcScores2}} function.
}
\examples{
example_vec <- c("a", "b", NA)
example_dat <- matrix(c("a", "a", "a",
                        "b", "b", "b",
                        NA, "b", "a"),
                      nrow = 3, byrow = T)
catDiffs2(example_vec, example_dat)
}

