% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-japan.R
\name{clean_japan}
\alias{clean_japan}
\title{Clean Japanese Leaders Downloaded from Baseball-Reference}
\usage{
clean_japan(dat, j_ids, type = c("bat", "pit"))
}
\arguments{
\item{dat}{data.frame.  Data scraped from Baseball-Reference}

\item{j_ids}{data.frame of japanese ids for players who have played in both
Japan and America.  The j_ids are custom, used only in this app}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting}
}
\value{
tbl_df of japanese statistics that have been cleaned.
}
\description{
Takes the Japanese data scraped from Baseball-Reference using
\code{\link{scrape_bref}} and cleans it so that it can be combined with
the data downloaded from PIA.  Only used to process data for the apps.
}
\examples{
curr_wd <- getwd()
setwd("N:/Apps/simScoresApp/data")
dat <- read.csv("0-downloads/bat_japan.csv", header = T, stringsAsFactors = F)
jids <- read.csv("manual-info/japan_ids.csv", header = T, stringsAsFactors = F)
y <- clean_japan(dat, jids, "bat")
setwd(curr_wd)
}

