% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{clean_PIA}
\alias{clean_PIA}
\title{Some Basic Cleaning on Data Downloaded from PIA}
\usage{
clean_PIA(minors, majors, okay_levels = c("High_A", "AA", "AAA", "MLB"))
}
\arguments{
\item{minors}{is a data.frame with the statistics from the minors that have
been download from PIA.}

\item{okay_levels}{the levels for which we wish to calculate MLEs.  Any
observations which aren't at these levels are removed.  Defaults to
c("High_A", "AA", "AAA", "MLB").}

\item{minors}{is a data.frame with the statistics from the minors that have
been download from PIA.}
}
\value{
\code{tbl_df} of cleaned statistics
}
\description{
This function changes column names of batting and pitching data downloaded
from PIA.  It also changes some of the labels, and removes rows where player
did not record a plate appearance.  Finally, it ensures that player played in
levels for which we have MLEs.
}
\details{
Changes column names, names of levels, and name of Florida
  organization to make typing easier.  For example,
  "Level.of.Play.Name.Abbrev" is changed to "Level."  Then removes any rows
  where the batter recorded 0 PA or the pitcher has 0 batters faced.
}
\examples{
curr_wd <- getwd()
setwd("N:/Apps/simScoresApp/data")
min <- read.csv("0-downloads/bat_minors2.csv", header = T, stringsAsFactors = F) \%>\% tbl_df()
maj <- read.csv("0-downloads/bat_majors2.csv", header = T, stringsAsFactors = F) \%>\% tbl_df()
x <- clean_PIA(min, maj)
x
setwd(curr_wd)
}

