% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{cleanNames}
\alias{cleanNames}
\title{Strips a name of non-alphabetic character when scraped from Baseball-Reference}
\usage{
cleanNames(dat)
}
\description{
Strips a name of non-alphabetic character when scraped from Baseball-Reference
}

