% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{combine}
\alias{combine}
\title{Combine Statistics with Bio, Position, and DL Info}
\usage{
combine(stats, bio, pos, dl)
}
\arguments{
\item{stats}{data.frame of player statistics.  Obtained from clean_PIA()}

\item{bio}{data.frame of Bio Data.  Must be in proper form.}

\item{dl}{data.frame of DL Time.  Must be in proper form.}

\item{position}{data.frame of Positions.  Must be in proper form.}
}
\value{
\code{tbl_df} of statistics that have been joined with \code{bio},
  \code{pos}, \code{dl}.
}
\description{
Takes the clean statistics (can be either actual or MLEs) and adds bio,
position, and dl info
}
\examples{
curr_wd <- getwd()
setwd("N:/Apps/simScoresApp/data")
stats <- read.csv("4-mles/bat-mles.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
bio <- read.csv("manual-info/bio_bat.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
pos <- read.csv("manual-info/positions.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
dl <- read.csv("manual-info/injuries.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
x <- combine(stats, bio, pos, dl)
setwd(curr_wd)
}

