% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{countWarnings}
\alias{countWarnings}
\title{Evaluates an Expression and Return the Number of Warnings}
\usage{
countWarnings(expr)
}
\arguments{
\item{expr}{the expression to evaluate}
}
\value{
list of length two. The first element of the list is names "ans" and
  it contains the results of the expression.  The second element of the list
  is named "warnings" and contains the number of warning generated by the
  expression.
}
\description{
Evaluates an expression.  Returns a list with the value of expression and the
number of warnings thrown by an expression.
}
\examples{
countWarnings(log(-1))
foo <- function(){
   warning("first warning!")
   warning("second warning!")
   warning("third warning!")
   invisible()
 }
 countWarnings(foo())
}
\references{
\url{http://stackoverflow.com/questions/4020239/warnings-does-not-work-within-a-function-how-can-one-work-around-this}
}

