% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{final_clean}
\alias{final_clean}
\title{Final Cleaning before Processing for Apps}
\usage{
final_clean(dat, type = c("bat", "pit"))
}
\arguments{
\item{dat}{data.frame obtained from \code{\link{combine}}.}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting}
}
\value{
\code{tbl_df} of statistics, bio, position, and DL information that
  are ready to be processed for the apps.
}
\description{
This removes extra columns that are no longer needed and sorts the data.
}
\examples{
curr_wd <- getwd()
setwd("N:/Apps/simScoresApp/data")
stats <- read.csv("4-mles/bat-mles.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
bio <- read.csv("manual-info/bio_bat.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
pos <- read.csv("manual-info/positions.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
dl <- read.csv("manual-info/injuries.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
x <- combine(stats, bio, pos, dl)
f <- final_clean(x, "bat")

# pitchers
pstats <- read.csv("4-mles/pit-mles.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
pbio <- read.csv("manual-info/bio_pit.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
px <- combine(pstats, pbio, pos, dl)
pf <- final_clean(px, "pit")
setwd(curr_wd)
}

