% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarity-scores.R
\name{findDiffs2}
\alias{findDiffs2}
\title{Calculates Differences for a Specified Player}
\usage{
findDiffs2(stats_fp, stats_comps, weights)
}
\arguments{
\item{stats_fp}{data.frame of one row.  The statistics for the focus player.}

\item{stats_comps}{data.frame with identical columns to \code{stats_fp}.  The
statistics for the group of players to be compared.  Each row corresponds
to the statistics for a player. Differences will be calculated by comparing
each row of \code{stats_comps} to \code{stats_fp}.}

\item{weights}{named vector of length \code{=ncol(stats_comps)} specifying a
weight for each variable.}
}
\value{
a vector of differences.  The jth element of the vector corresponds
  to the difference between \code{stats_fp} and the jth row of
  \code{stats_comps}.  Differences are calculated using Gower's distance
  formula.
}
\description{
Compares statistics for a player of interest to a the statistics for a large
group of players.  Returns the distance from each player in the group to the
player of interest.
}
\details{
\code{findDiffs2} separates the statistics into quantitative and
  categorical statistics.  It then passes the categorical statistics (for
  both the focus player and the comps) and the categorical weights to
  \code{\link{catDiffs2}} to calculate differences for categorical variables.
  It passes the quantitative statistics and weights to
  \code{\link{numDiffs2}} calculate differences for quantitative variables.

  Using Gower's formula, distances for categorical variables between player
  \eqn{j} and the focus player are 0 if the players have the same value of the
  variable, and 1 if it's different.  Difference for the \eqn{i^{th}}
  quantitative variable for the \eqn{j^{th}} player are \eqn{d_{ij} = (stat_{ij} -
  stat_{i,fp}) / range(stat_i)}.

  \code{findDiffs2} then records missing data as \eqn{m_{ij} = 0} if the \eqn{i^{th}}
  statistic for player \eqn{j} or for the focus player is missing.  Let
  \eqn{delta_{j}} be the distance between player \eqn{j} and the focus player and
  \eqn{w_i} be the weight for the \eqn{i^{th}} statistic (\eqn{i = 1...I}).  Using
  Gower's distance metric, \deqn{delta_j = sum(i=1:I; w_i m_{ij} d_{ij}) /
  sum(i=1:I; w_i d_{ij})}
}
\examples{
### Using the iris data
wts <- (1:5) \%>\% setNames(names(iris))
fp <- data.frame(5,4,3,2,"setosa") \%>\% setNames(names(iris))
i <- head(iris)
findDiffs2(fp, i, wts)
# ratio of weights are all that matters
wts <- wts / 10
findDiffs2(fp, i, wts) # same thing!
# columns not names in wts will be removed
wts <- wts[-4]
findDiffs2(fp, i, wts) # still works

# can handle missing data!
i <- head(iris)
i[3,4] <- NA
findDiffs2(fp, i, wts)
}
\references{
Gower, J. C. (1971) A general coefficient of similarity and some
  of its properties, Biometrics 27, 857 - 874.
}

