% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_curve_app_data}
\alias{get_curve_app_data}
\title{Process Statistics for Aging Curve Apps}
\usage{
get_curve_app_data(app_data_dir, app_dir, type = c("bat", "pit"))
}
\arguments{
\item{app_dir}{character.  Path to the curve app that is being used}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting}

\item{dat_dir}{character.  Path to data cleaned by \code{\link{get_MLEs}}.}
}
\value{
nothing.  Just saves the processed data
}
\description{
Creats data for the aging curve apps
}
\examples{
get_curve_app_data(app_data_dir = "N:/Apps/simScoresApp/data/6-for-apps/batters/",
                   app_dir = "N:/Apps/simScoresApp/b-comparison-app/",
                   type = "bat")
get_curve_app_data(app_data_dir = "N:/Apps/simScoresApp/data/6-for-apps/pitchers/",
                   app_dir = "N:/Apps/simScoresApp/p-curve-app/",
                   type = "pit")
}

