% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_MLEs}
\alias{get_MLEs}
\title{Clean Stats and Get MLEs}
\usage{
get_MLEs(minors, majors, bio_info, position, injuries, park_factors,
  year_factors, multipliers, ignore_players, japan_ids, download_dir, clean_dir,
  pf_dir, yf_dir, mle_dir, miss_dir, app_dir, type = c("bat", "pit"),
  wd = NULL, yrs = 2005:2015)
}
\arguments{
\item{minors}{character.  Path to csv file of minor league data.  Used in
\code{\link{read.csv}}.  Must be downloaded from PIA in proper format}

\item{majors}{character.  Path to csv file of major league data.  Used in
\code{\link{read.csv}}.  Must be downloaded from PIA in proper format}

\item{bio_info}{character.  Path to csv file of bio data.  Used in
\code{\link{read.csv}}.  Must be in proper format}

\item{position}{character.  Path to csv file of position data.  Used in
\code{\link{read.csv}}.  Must be in proper format}

\item{injuries}{character.  Path to csv file of DL time.  Used in
\code{\link{read.csv}}. Must be in proper format}

\item{park_factors}{character.  Path to csv file of park factors.  Used in
\code{\link{read.csv}}.  Must be in proper format}

\item{year_factors}{character.  Path to csv file of year factors.  Used in
\code{\link{read.csv}}.  Must be in proper format}

\item{multipliers}{character.  Path to csv file of multipliers for MLEs. Used
in \code{\link{read.csv}}.  Must be in proper format}

\item{ignore_players}{character.  Path to csv file of players to ignore. Used
in \code{\link{read.csv}}.  Must be in proper format}

\item{clean_dir}{character.  Path to write csv file of cleaned data.  Used in
\code{\link{write.csv}}.}

\item{pf_dir}{character.  Path to write csv file of stats adjusted for park
factors.  Used in \code{\link{write.csv}}.}

\item{yf_dir}{character.  Path to write csv file of stats adjusted for year
effects  Used in \code{\link{write.csv}}.}

\item{mle_dir}{character.  Path to write csv file of MLEs.  Used in
\code{\link{write.csv}}.}

\item{miss_dir}{character.  Path to write csv file with missing data.  Used
in \code{\link{write.csv}}.}

\item{app_dir}{character.  Path to write csv file with data for apps.  Used
in \code{\link{write.csv}}.}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting}

\item{wd}{character.  Used in \code{\link{setwd}} to change the working
directory}
}
\value{
nothing.  Saves the data into the specified directories.
}
\description{
This function is a wrapper which cleans data, then adjusts for park factors,
then calculates MLEs, then adds biographical, position, and DL information.
}
\examples{
# batters
get_MLEs(minors = "0-downloads/bat_minors2.csv",
         majors = "0-downloads/bat_majors2.csv",
         bio_info = "manual-info/bio_bat.csv",
         position = "manual-info/positions.csv",
         injuries = "manual-info/injuries.csv",
         park_factors = "manual-info/Park_Factors.csv",
         year_factors = "manual-info/Year-Factors.csv",
         multipliers = "manual-info/Level_Multipliers.csv",
         ignore_players = "manual-info/players_to_ignore.csv",
         japan_ids = "manual-info/japan_ids.csv",
         download_dir = "0-downloads/",
         clean_dir = "1-cleaned/batters/",
         pf_dir = "2-park-adjusted/",
         yf_dir = "3-year-adjusted/",
         mle_dir = "4-mles/",
         miss_dir = "5-missing/",
         app_dir = "6-for-apps/batters/",
         wd = "N:/Apps/simScoresApp/data/",
         type = "bat",
         yrs = 2005:2014)
# pitchers
get_MLEs(minors = "0-downloads/pit_minors2.csv",
         majors = "0-downloads/pit_majors2.csv",
         bio_info = "manual-info/bio_pit.csv",
         position = "manual-info/positions.csv",
         injuries = "manual-info/injuries.csv",
         park_factors = "manual-info/Park_Factors.csv",
         year_factors = "manual-info/Year-Factors.csv",
         multipliers = "manual-info/Level_Multipliers.csv",
         ignore_players = "manual-info/players_to_ignore.csv",
         japan_ids = "manual-info/japan_ids.csv",
         download_dir = "0-downloads/",
         clean_dir = "1-cleaned/pitchers/",
         pf_dir = "2-park-adjusted/",
         yf_dir = "3-year-adjusted/",
         mle_dir = "4-mles/",
         miss_dir = "5-missing/",
         app_dir = "6-for-apps/pitchers/",
         wd = "N:/Apps/simScoresApp/data/",
         type = "pit",
         yrs = 2005:2014)
}

