% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getLastF}
\alias{getLastF}
\title{Shorten Name}
\usage{
getLastF(display_name)
}
\arguments{
\item{display_name}{character in the format "Last; First"}
}
\value{
character in the format "LastF"
}
\description{
Takes a Display Name from the apps and turns it into "LastF"
}
\examples{
getLastF("Schmidt; Mike J.")
}

