% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{identify_missing}
\alias{identify_missing}
\title{Find Missing Data}
\usage{
identify_missing(dat, ignore)
}
\arguments{
\item{dat}{data.frame.  This contains both statistics and bio data so that
missing data may be identified by the user.  Obtained from
\code{\link{combine}}.}

\item{ignore}{data.frame.  This contains MLBID that the user wants to ignore
when identifying missing data.  Must be a data.frame with a column named
MLBID}
}
\value{
a data.frame containing bio info for any player with missing data.
  Most of the time, it should just be bio data that is missing.  If however,
  a batting/pitching statistic is missing, the player with the missing data
  will be identified.
}
\description{
Identifies players with missing data.  Records these players and saves files
with these players' MLBID so that missing data may be filled in by the user
at a later time.
}
\examples{
curr_wd <- getwd()
setwd("N:/Apps/simScoresApp/data")
stats <- read.csv("4-mles/bat-mles.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
bio <- read.csv("manual-info/bio_bat.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
pos <- read.csv("manual-info/positions.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
dl <- read.csv("manual-info/injuries.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
ignore <- read.csv("manual-info/players_to_ignore.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
x <- combine(stats, bio, pos, dl)
m <- identify_missing(x, ignore)
setwd(curr_wd)
}

