% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-data.R
\name{identify_missing_positions}
\alias{identify_missing_positions}
\title{Find Missing Position Information}
\usage{
identify_missing_positions(dat, ignore)
}
\arguments{
\item{dat}{data.frame containing MLBID, Year, and Position columns.  Obtained
from \code{\link{combine}}.}

\item{ignore}{data.frame.  This contains MLBID that the user wants to ignore
when identifying missing data.  Must be a data.frame with a column named
MLBID}
}
\value{
a data.frame containing the MLBID and Year in which the Player's
  Position is missing.
}
\description{
Identifies player-seasons for which the player's position is missing
}
\examples{
curr_wd <- getwd()
setwd("N:/Apps/simScoresApp/data")
stats <- read.csv("4-mles/bat-mles.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
bio <- read.csv("manual-info/bio_bat.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
pos <- read.csv("manual-info/positions.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
dl <- read.csv("manual-info/injuries.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
ignore <- read.csv("manual-info/players_to_ignore.csv", header = T, stringsAsFactors = F) \%>\% tbl_df() \%>\% mutate(MLBID = as.character(MLBID))
x <- combine(stats, bio, pos, dl)
m2 <- identify_missing_positions(x, ignore)
setwd(curr_wd)
}

