% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/similarity-scores.R
\name{numDiffs2}
\alias{numDiffs2}
\title{Differences in Numeric Variables}
\usage{
numDiffs2(vec, dat)
}
\arguments{
\item{vec}{a numeric vector.  Each value in the vector corresponds to a
unique variable.}

\item{dat}{a numeric matrix or data.frame of numeric variables.  The number
of columns must be equal to the length of \code{vec}.}
}
\value{
a matrix of the same dimensions as \code{dat}.  The ith entry in the
  jth row corresponds to the absolute value of the difference between the jth
  value of \code{vec} and the jth value of the ith row of \code{dat}.If
  either the ith value of the jth row of \code{dat} or the jth value of
  \code{vec} are missing, and NA will be returned.
}
\description{
Computes the scaled difference between a vector of numeric variables and a
matrix of numeric variables.  Should not usually be called by hand.  Instead,
use the \code{\link{calcScores2}} function.
}
\examples{
example_vec <- c(1,2, NA)
example_dat <- iris[1:6, 1:3]
numDiffs2(example_vec, example_dat)
}

