% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/age-curves.R
\name{pred_curve}
\alias{pred_curve}
\title{Predict Aging Curve}
\usage{
pred_curve(ages, coefs, poly_model)
}
\arguments{
\item{ages}{numeric.  The ages for which to estimate the player's aging
curve.}

\item{coefs}{The coefficients from the fitted model.  Can be either a vector
or a matrix where each column consists of a different set of coefficients.}

\item{poly_model}{the model used to decompose the \code{ages} vector into
orthogonal polynomials.  Must have degree equal to the number of rows in
the \code{coefs} matrix or, if \code{coefs} is a vector, the length of
\code{coefs}.}
}
\value{
data.frame containing the ages and the estimated aging curve for each
  age.
}
\description{
Predicts the aging curve for a sequence of ages using the fixed effect
coefficients from a fitted model
}

