% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean-japan.R
\name{scrape_bref}
\alias{scrape_bref}
\title{Scrapes Batting or Pitching Leaders from Baseball-Reference}
\usage{
scrape_bref(yrs, type = c("bat", "pit"),
  brefBaseURL = "http://www.baseball-reference.com/japan/leader.cgi?type=bat&year=")
}
\arguments{
\item{yrs}{numeric. Vector of the years to scrape the data from}

\item{type}{Whether to scrape batting or pitching data.  Defaults to batting}

\item{bref_base_url}{character containing the url with the correct
Baseball-Reference leaderboard.  Defaults to Japanese leaders.}
}
\value{
\code{tbl_df} of data scraped from Baseball-Reference.
}
\description{
Takes the url for a Baseball-References leaderboard and scrapes the data over
a specified timeframe
}
\details{
Pastes \code{yrs} to the end of \code{brefBaseURL} to create a
  vectors of urls.  Then scrapes the first table at each of these urls using
  \code{\link{readHTMLTable}}.  Then binds the table for each year using
  \code{\link{rbind_all}} and returns these stacked tables as a
  \code{\link{tbl_df}} object.
}
\examples{
# batters
x <- scrape_bref(2005:2014)
# pitchers
y <- scrape_bref(2005:2014, type = "pit")
}

