% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-processing.R
\name{select_advanced}
\alias{select_advanced}
\title{Selects Advanced Statistics}
\usage{
select_advanced(dat, type = c("bat", "pit"))
}
\arguments{
\item{dat}{data.frame of statistics.  Obtained from
\code{\link{final_clean}}.}
}
\value{
data.frame containing a player's advanced stats
}
\description{
Extracts just the advanced statistics from the specified data.frame
}
\examples{
x <- read.csv("N:/Apps/simScoresApp/data/6-for-apps/batters/bat-raw.csv", header = T, stringsAsFactors = F) \%>\% tbl_df()
x1 <- select_advanced(x, type = "bat")
}

