% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-processing.R
\name{select_traditional}
\alias{select_traditional}
\title{Selects Traditional Statistics}
\usage{
select_traditional(dat)
}
\arguments{
\item{dat}{data.frame of statistics.  Obtained from
\code{\link{final_clean}}.}
}
\value{
data.frame containing a player's traditional stats
}
\description{
Extracts just the traditional statistics from the specified data.frame
}
\examples{
x <- read.csv("N:/Apps/simScoresApp/data/6-for-apps/batters/bat-raw.csv", header = T, stringsAsFactors = F) \%>\% tbl_df()
x1 <- select_traditional(x)
}

