% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{substr_right}
\alias{substr_right}
\title{Right Substring of a Character Vector}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{a character vector}

\item{n}{number of characters}
}
\value{
character.  The right n characters of the string
}
\description{
Extract the right n characters from a string
}
\examples{
substr_right("some text in a string", 6)
substr_right("some text in a string", 8)
}

