% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-processing.R
\name{sum_stats}
\alias{sum_stats}
\title{Sum Statistics over Grouping Variables}
\usage{
sum_stats(dat, type = c("bat", "pit"), group_level = F)
}
\arguments{
\item{dat}{data.frame of statistics.  Obtained from
\code{\link{final_clean}}.}

\item{type}{character.  Whether these are batting or pitching data.  Defaults
to batting}

\item{group_level}{boolean.  Whether these statistics should be grouped by
level.  FALSE for use in \code{\link{age_comps}}, TRUE for use in
\code{\link{level_comps}}.Defaults to FALSE.}
}
\value{
data.frame containing a player's statistics that have been summed. If
  \code{group_level} is TRUE, these statistics will be summed for each level
  a player recorded statistics.  If it is FALSE, it will be summed by age.
}
\description{
Sums statistics for batters or pitchers.  Statistics can then be used in
either \code{\link{age_comps}} or \code{\link{level_comps}}.
}
\examples{
x <- read.csv("N:/Apps/simScoresApp/data/6-for-apps/batters/bat-raw.csv", header = T, stringsAsFactors = F) \%>\% tbl_df()
x1 <- sum_stats(x)
}

