% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app-processing.R
\name{switch_names}
\alias{switch_names}
\title{Change Names for Presentation}
\usage{
switch_names(dat, names_to_switch = c("X1B", "X2B", "X3B", "ML.Org",
  "DL.Days", "PA.G", "IP.G", "K.perc", "BB.perc", "K.9", "BB.9", "HR.9"))
}
\arguments{
\item{dat}{data.frame}

\item{names_to_switch}{}
}
\value{
data.frame with names changed for display
}
\description{
Changes the names of certain statistics to make them look nicer for
presentation
}
\examples{
x <- read.csv("N:/Apps/simScoresApp/data/6-for-apps/batters/bat-raw.csv", header = T, stringsAsFactors = F) \%>\% tbl_df()
x1 <- switch_names(x)
}

