% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{update_apps}
\alias{update_apps}
\title{Updates Data and Deploys Apps with Updated Data to ShinyApps}
\usage{
update_apps(wd = NULL, yrs = 2005:2015, ...)
}
\arguments{
\item{wd}{character.  The location of the working directory where the apps
are located.}

\item{yrs}{numeric.  The years for which to scrape Baseball-Reference data
for Japan.}
}
\value{
nothing.  Just downloads and processes data then uploads the
  processed data to shinyapps.io
}
\description{
The use of this function is described in the "How the Updates Work" vignette.
It downloads data from Japan, then cleans Japanese and PIA data.  It then
performs park, year, and level adjustments on the statistics.  Missing data
are recorded, then the data are processed for the apps.  The data are saved
into the appropriate folders for each app and then the updated apps are
deployed to shinyapps.io
}
\examples{
update_apps(wd = "N:/Apps/simScoresApp")
}

