% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{year_diff}
\alias{year_diff}
\title{Time between Two Dates}
\usage{
year_diff(d1, d2)
}
\arguments{
\item{d1}{first date in character form.  Must be in mdy form for lubridate}

\item{d2}{second date in character form.  Must be in mdy form for lubridate}
}
\value{
number of years between \code{d1} and \code{d2}
}
\description{
Find the number of years between two dates
}
\examples{
year_diff("9/13/1990", "12/5/2014")
}

